pyproject.toml
[tool.poetry]
name = "habit-tracker-cw7"
version = "0.1.0"
description = ""
authors = ["pozdniakova olga <caramelca07@mail.ru>"]
readme = "README.md"
packages = [{include = "habit_tracker_cw7"}]

[tool.poetry.dependencies]
python = "^3.11"
poetry = "^1.6.1"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
